refactor(requestoverview): shift date time into utils
