+++ src/utils/date.ts
@@ -1,0 +1,20 @@
+interface DisplayedDateTime {
+  date: string
+  time: string
+}
+
+export const getDateTimeFromUnixTime = (
+  unixTime: number
+): DisplayedDateTime => {
+  const date = new Date(unixTime)
+  return {
+    date: date.toLocaleDateString("en-GB", {
+      month: "short",
+      year: "numeric",
+      day: "numeric",
+    }),
+    time: date.toLocaleTimeString("en-GB", {
+      timeStyle: "short",
+    }),
+  }
+}
