+++ src/utils/index.ts
@@ -9,3 +9,4 @@ export * from "./toasts"
 export * from "./validators"
 export * from "./legacy"
 export * from "./text"
+export * from "./date"
